move visual files to UI folder
