+++ src/App.tsx
@@ -1,4 +1,4 @@
-import Routes from './components/Routes.tsx';
+import Routes from '~@/Routes.tsx';
 import styled from 'styled-components';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
