+++ src/ui/components/ManageProductMenu/Tabs/ProductEditForm.tsx
@@ -5,7 +5,7 @@ import { MdCheckBox, MdCheckBoxOutlineBlank, MdEuro } from 'react-icons/md';
 import styled from 'styled-components';
 import type { ManageProductType } from '@Types/ManageProductType.ts';
 
-import { theme } from '~@/theme';
+import { theme } from '~@/ui/theme';
 import { useEditProductForm } from '@Hooks/form/useEditProductForm.ts';
 import { IoMdAddCircle, IoMdRemoveCircle } from 'react-icons/io';
 import { FiPackage } from 'react-icons/fi';
