+++ src/ui/components/Navbar/TurnAdminModeButton.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { theme } from '~@/theme';
+import { theme } from '~@/ui/theme';
 
 type TurnAdminModeButtonProps = {
   active: boolean;
