+++ src/ui/components/OrderMenu/OutOfStock/ItemCardOutOfStock.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { theme } from '~@/theme';
+import { theme } from '~@/ui/theme';
 
 export const ItemCardOutOfStock = () => {
   return (
